requirements.txt
alembic==1.2.1
autopep8==1.4.4
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
google-api-core==1.10.0
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.15.0
google-cloud-texttospeech==0.4.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
Mako==1.1.0
MarkupSafe==1.1.1
protobuf==3.7.1
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pydub==0.23.1
pymongo==3.8.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
regex==2019.4.14
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
urllib3==1.24.2
Werkzeug==0.15.2
WTForms==2.2.1
